+++ jugaad_data/__init__.py
@@ -1,3 +1,3 @@
-__version__ = "0.17"
+__version__ = "0.19"
